Refactor: permissionService, more clean statusService

+++ src/main/java/com/lmt/Kanban/service/impl/StatusServiceImpl.java
@@ -1,10 +1,8 @@
 package com.lmt.Kanban.service.impl;
 
-import ch.qos.logback.core.status.StatusUtil;
 import com.lmt.Kanban.dto.request.CreateStatusRequest;
 import com.lmt.Kanban.dto.request.UpdateStatusRequest;
 import com.lmt.Kanban.dto.response.StatusResponse;
-import com.lmt.Kanban.dto.response.TaskResponse;
 import com.lmt.Kanban.entity.Board;
 import com.lmt.Kanban.entity.Status;
 import com.lmt.Kanban.exception.ErrorCode;
@@ -15,7 +13,7 @@ import com.lmt.Kanban.repository.BoardRepository;
 import com.lmt.Kanban.repository.StatusRepository;
 import com.lmt.Kanban.service.BoardService;
 import com.lmt.Kanban.service.StatusService;
-import com.lmt.Kanban.service.TaskPermissionService;
+import com.lmt.Kanban.service.PermissionService;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -25,8 +23,7 @@ import java.util.List;
 @Service
 @AllArgsConstructor
 public class StatusServiceImpl implements StatusService {
-    private final BoardRepository boardRepository;
-    private final TaskPermissionService taskPermissionService;
+    private final PermissionService permissionService;
     private final StatusRepository statusRepository;
     private final BoardService boardService;
 
@@ -34,10 +31,9 @@ public class StatusServiceImpl implements StatusService {
 
     @Override
     public StatusResponse createStatus( Board board,CreateStatusRequest request) {
+        permissionService.checkBoardAdminOrOwner(board.getId());
+
         Status status = statusMapper.toEntity(request);
-
-        //TODO kiểm tra Permission
-
         status.setBoard(board);
 
         Integer maxPosition = statusRepository.findMaxPositionByBoardId(request.getBoardId());
@@ -49,25 +45,26 @@ public class StatusServiceImpl implements StatusService {
     }
 
     @Override
-    public StatusResponse updateStatus(UpdateStatusRequest request) {
-        Status status = getStatusEntity(request.getId());
+    public StatusResponse updateStatus(Long id,UpdateStatusRequest request) {
+        Status status = getStatusEntity(id);
+
+        permissionService.checkBoardAdminOrOwner(status.getBoard().getId());
 
         statusMapper.updateStatusFromRequest(request, status);
-        return createStatusResponse(status);
+        return statusMapper.toResponse(status);
     }
 
     @Override
     @Transactional
     public void movePosition(Long statusId, Integer newPosition) {
-        Status status = statusRepository.findById(statusId)
-                .orElseThrow(() -> new ResourceNotFoundException(ErrorCode.STATUS_NOT_FOUND,"Status not found with ID: " + statusId));
-
+        Status status = getStatusEntity(statusId);
         Long boardId = status.getBoard().getId();
         Integer oldPosition = status.getPosition();
 
-        if (oldPosition.equals(newPosition)) {
-            return;
-        }
+        permissionService.checkBoardAdminOrOwner(boardId);
+
+        if (oldPosition.equals(newPosition)) return;
+
 
         if (newPosition < oldPosition) {
             statusRepository.shiftRight(boardId, newPosition, oldPosition);
@@ -86,8 +83,10 @@ public class StatusServiceImpl implements StatusService {
         Long boardId = status.getBoard().getId();
         Integer position = status.getPosition();
 
+        permissionService.checkBoardAdminOrOwner(boardId);
+
         statusRepository.delete(status);
-//        statusRepository.shiftLeftAfterDelete(boardId, position);
+        statusRepository.shiftLeftAfterDelete(boardId, position);
     }
 
     @Override
@@ -103,29 +102,31 @@ public class StatusServiceImpl implements StatusService {
     @Override
     public StatusResponse getStatusById(Long statusId) {
         Status status = getStatusEntity(statusId);
-        return createStatusResponse(status);
+
+        permissionService.checkBoardMember(status.getBoard().getId());
+        return statusMapper.toResponse(status);
     }
 
     @Override
     public List<StatusResponse> getAllStatus(Long boardId) {
-        if(!boardRepository.existsById(boardId)){
-            throw new ResourceNotFoundException(ErrorCode.BOARD_NOT_FOUND,"Board not found with ID: " + boardId);
-        }
+        boardService.validateBoard(boardId);
+
+        permissionService.checkBoardMember(boardId);
 
         List<Status> statusList = statusRepository.findByBoardId(boardId);
 
         return statusList.stream()
-                .map(this::createStatusResponse)
+                .map(statusMapper::toResponse)
                 .toList();
     }
 
-    @Override
-    public List<StatusResponse> getAllStatus() {
-        List<Status> statusList = statusRepository.findAll();
-        return statusList.stream()
-                .map(this::createStatusResponse)
-                .toList();
-    }
+//    @Override
+//    public List<StatusResponse> getAllStatus() {
+//        List<Status> statusList = statusRepository.findAll();
+//        return statusList.stream()
+//                .map(statusMapper::toResponse)
+//                .toList();
+//    }
 
     @Override
     public Boolean checkStatusInBoard(Long boardId, Long statusId) {
@@ -133,18 +134,6 @@ public class StatusServiceImpl implements StatusService {
     }
 
     private Status getStatusEntity (Long statusId) {
-        System.out.println(statusId);
         return statusRepository.findById(statusId).orElseThrow(() -> new ResourceNotFoundException(ErrorCode.STATUS_NOT_FOUND,"Status not found with ID: " + statusId));
     }
-
-    private StatusResponse createStatusResponse(Status status) {
-        return StatusResponse.builder()
-                .id(status.getId())
-                .color(status.getColor())
-                .name(status.getName())
-                .position(status.getPosition())
-                .boardId(status.getBoard().getId())
-                .build();
-    }
-
 }
